Moved up to panel ctrl to make it more generic
